requirements.txt
attrs==17.4.0
bencode.py==2.0.0
coverage==4.5.1
flake8==3.5.0
mccabe==0.6.1
mypy==0.560
pep8==1.7.1
pluggy==0.6.0
psutil==5.4.3
py==1.5.2
pycodestyle==2.3.1
pycrypto==2.6.1
pyflakes==1.6.0
pytest==3.4.1
six==1.11.0
tox==2.9.1
typed-ast==1.1.0
virtualenv==15.1.0
